backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
sqlalchemy==2.0.22
pyjwt==2.8.0
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0